requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.20
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
Werkzeug==2.3.7
